+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Note the first digit of every adapter version corresponds to the major version o
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
 ### 4.10.5.6.1
+- If TCFv2 string is present, it will now be added when setting GDPR consent.
 - This version of the adapters has been certified with InMobiSDK 10.5.6.
 
 ### 4.10.5.6.0